libthinkpad: 2.0 - Remove the X11 API
Why? Well I implemented that API in order to create
a new UI application that can manage screens.

As I progressed the API I realized that that configuring
X is heavy and tedious work, and is really error prone.

And the X11 configuration system has nothing to do in a
hardware library.

Maybe some other day.